DB methods refactor. Created separate class for it.
